+++ src/nrepl/connect-types.ts
@@ -1,0 +1,4 @@
+export enum ConnectType {
+  Connect,
+  JackIn,
+}