+++ src/state.ts
@@ -6,6 +6,8 @@ import * as path from 'path';
 import * as child from 'child_process';
 import { getStateValue, setStateValue } from '../out/cljs-lib/cljs-lib';
 import * as projectRoot from './project-root';
+import { getCustomConnectSequences, ReplConnectSequence } from './nrepl/connectSequence';
+import { ConnectType } from './nrepl/connect-types';
 
 let extensionContext: vscode.ExtensionContext;
 export function setExtensionContext(context: vscode.ExtensionContext) {
@@ -189,16 +191,37 @@ function getProjectWsFolder(): vscode.WorkspaceFolder | undefined {
 /**
  * Figures out the current clojure project root, and stores it in Calva state
  */
-export async function initProjectDir() {
+export async function initProjectDir(
+  connectType: ConnectType,
+  connectSequence: ReplConnectSequence,
+  disableAutoSelect = false
+) {
   const candidatePaths = await projectRoot.findProjectRoots();
   const active_uri = vscode.window.activeTextEditor?.document.uri;
   const closestRootPath: vscode.Uri = active_uri
     ? projectRoot.findClosestParent(active_uri, candidatePaths)
     : undefined;
-  const projectRootPath: vscode.Uri = await projectRoot.pickProjectRoot(
-    candidatePaths,
-    closestRootPath
-  );
+
+  const sequences: ReplConnectSequence[] = getCustomConnectSequences();
+
+  const defaultSequence = connectSequence
+    ? connectSequence
+    : disableAutoSelect
+    ? undefined
+    : sequences.find(
+        (s) =>
+          (connectType === ConnectType.Connect ? s.autoSelectForConnect : s.autoSelectForJackIn) &&
+          !!s.projectRootPath
+      );
+
+  const projectRootPath: vscode.Uri = defaultSequence
+    ? vscode.Uri.parse(
+        path.resolve(
+          vscode.workspace.workspaceFolders[0].uri.fsPath,
+          ...(defaultSequence?.projectRootPath ? defaultSequence.projectRootPath : [])
+        )
+      )
+    : await projectRoot.pickProjectRoot(candidatePaths, closestRootPath, connectType);
   if (projectRootPath) {
     setStateValue(PROJECT_DIR_KEY, projectRootPath.fsPath);
     setStateValue(PROJECT_DIR_URI_KEY, projectRootPath);